Welcome UI: Set P3A opt-in from a feature flag.
Define a base::Feature for enabling the P3A opt-in user elements
for the study and propagate this value as a boolean into the
welcome ui content page.

The new opt-in element can be shown  by launching brave with

    npm start -- --enable-features=P3AOptIn

on a fresh profile, or visiting brave://welcome later.

This disables the kP3AEnable pref if it's still in its default
value. The FeatureList singleton isn't available at the
time the brave_p3a_service::RegisterPrefs is called, so we can't
gate the default value on the feature and must set it when the
welcome ui runs.

The feature flag adds a Checkbox element to the Welcome UI for
the opt-in study, converting the `ShieldsBox` to a function so
a useState hook can maintain its state locally. Similarly, apply
a useDispatch hook to trigger a custom action posting checkbox
state changes to the app state for reporting.

Thanks to @DJAndries for help with redux.

The final Checkbox state is reported with other P3A measurements
when the 'Done' button is clicked at the end of the workflow as
part of explicit confirmation.

TODO: Probably we should set the initial checkbox to the value of
the pref when we open the page so re-launching the welcome ui doesn't
end up recording a disable.