P3A: Move metric list to a separate file.
Adding new P3A probe requires updating the list of allowed metrics
so the new measurement can be reported. This is a lower review bar
than changing the reporting code itself, so we'd like to have it
in a different place so we can have a separate list of code owners
for it.

Use an include guard to prevent accidental duplicates.

Resolves https://github.com/brave/brave-browser/issues/22263